Update flutter_facebook_auth.podspec
I had a bunch of errors about the FBSDKCoreKit needing a deployment target of iOS 11 (I had 11 set in my cocoapods due to other dependencies) and was unable to build.

This fixed it.

Not sure if this is a viable fix - im in a no way a native dev, but leaving it here in case its necessary / helpful. If not feel free to close it out.